# Conflicts:
#	node_modules/.cache/.eslintcache
#	src/components/Common/Header.js
#	src/components/Common/SideBar.js
#	src/components/MyPage/MyPage.js
#	src/components/StoreItemDailyPrice/StoreItemDailyPrice.js
#	src/components/utility/common.js

+++ src/components/utility/common.js
@@ -22,57 +22,58 @@ export function isNotNullOrNonEmpty(value) {
 
 /* 토큰 만료 체크 */
 export function isTokenExpired(token) {
-    if (!isNullUndefinedOrEmpty(token)) {
-        return true;
-    }
-    const expirationTime = jwtDecode(token).exp * 1000;
-    const currentDate = Date.now();
-    return currentDate > expirationTime; // 토큰 만료 true 반환
+  if (!isNullUndefinedOrEmpty(token)) {
+    return true;
+  }
+  const expirationTime = jwtDecode(token).exp * 1000;
+  const currentDate = Date.now();
+  return currentDate > expirationTime; // 토큰 만료 true 반환
 }
+
 
 
 export function isNullUndefinedOrEmpty(value) {
-    const valueStr = value + '';
-    if (valueStr === 'null' || value === 'undefined' || /^\s*$/.test(valueStr)) {
-        return null;
-    }
-    return value;
+  const valueStr = value + '';
+  if (valueStr === 'null' || value === 'undefined' || /^\s*$/.test(valueStr)) {
+    return null;
+  }
+  return value;
 }
 
 
-export function redirectLogin() {
-    alert('세션이 유효하지 않아 로그인이 필요합니다!');
-    window.location.href = '/login';
+export function refreshTokenInvalid() {
+  alert('세션이 만료되어 재로그인이 필요합니다!');
+  window.location.href = '/login';
 }
 
 export function accessTokenRefresh() {
-    const refreshToken = localStorage.getItem('refreshToken');
-    console.log('accessTokenRefresh() => refreshToken:', refreshToken);
+  const refreshToken = localStorage.getItem('refreshToken');
+  console.log('accessTokenRefresh() => refreshToken:', refreshToken);
 
-    if (isTokenExpired(refreshToken)) {
-        redirectLogin();
-    }
+  if (!isNullUndefinedOrEmpty(refreshToken)) {
+    refreshTokenInvalid();
+  }
 
-    // 서버로 리프레시 토큰을 사용하여 새 액세스 토큰을 요청
-    axios.post('/api/member/accTokRefresh', {refreshToken})
-        .then(response => {
-            console.log('/api/member/accTokRefresh => response: ', response);
-            const newAccessToken = response.data.accessToken;
-            if (!isNullUndefinedOrEmpty(newAccessToken)) {
-                redirectLogin();
-                return;
-            }
-            sessionStorage.setItem('accessToken', newAccessToken);
-            console.log('refreshToken으로 accessToken 재발급 완료! =>');
-        })
-        .catch(error => {
-            redirectLogin();
-        });
+  // 서버로 리프레시 토큰을 사용하여 새 액세스 토큰을 요청
+  axios.post('/api/member/accTokRefresh', { refreshToken })
+    .then(response => {
+      console.log('/api/member/accTokRefresh => response: ', response);
+      const newAccessToken = response.data.accessToken;
+      if (!isNullUndefinedOrEmpty(newAccessToken)) {
+        refreshTokenInvalid();
+        return;
+      }
+      sessionStorage.setItem('accessToken', newAccessToken);
+      console.log('refreshToken으로 accessToken 재발급 완료! =>');
+    })
+    .catch(error => {
+      refreshTokenInvalid();
+    });
 }
 
 
 export function logout() {
-    const isConfirmed = window.confirm('정말로 로그아웃 하시겠습니까?');
+  const isConfirmed = window.confirm('정말로 로그아웃 하시겠습니까?');
 
     if (isConfirmed) {
         sessionStorage.clear();
@@ -83,9 +84,9 @@ export function logout() {
 }
 
 export function getAccessToken(encodingOrDecodingType) {
-    encodingOrDecodingType = encodingOrDecodingType.toLowerCase();
+  encodingOrDecodingType = encodingOrDecodingType.toLowerCase();
 
-    const accessToken = sessionStorage.getItem('accessToken');
+  const accessToken = sessionStorage.getItem('accessToken');
 
     if (!isTokenExpired(accessToken)) { // 엑세스 토큰이 유효하면
         if (encodingOrDecodingType === 'a') {
@@ -139,99 +140,98 @@ export function sendAxiosRequest(url, method, data, successCallback, errorCallba
         };
     }
 
-    console.log('sendAxiosRequest().axiosConfig: ', axiosConfig);
+  console.log('sendAxiosRequest().axiosConfig: ', axiosConfig);
 
-    axios(axiosConfig)
-        .then(successCallback)
-        .catch(errorCallback);
+  axios(axiosConfig)
+    .then(successCallback)
+    .catch(errorCallback);
 }
 
 
-export function sendAxiosMultipartRequest(url, formData, successCallback, errorCallback,  jwtToken) {
-    console.log('sendAxiosMultipartRequest의 요청 URL: ', url);
-    console.log('sendAxiosMultipartRequest의 요청 데이터: ', formData);
-    const axiosConfig = {
-        timeout: 5000,
-        url: url,
-        method: 'POST',
-        data: formData,
-        headers: {
-            'Content-Type': 'multipart/form-data',
-            'Authorization': jwtToken !== 'UA' ? `Bearer ${jwtToken}` : 'UA',
-        }
-    };
-    axios(axiosConfig).then(successCallback).catch(errorCallback);
+export function sendAxiosMultipartRequest(url, formData, successCallback, errorCallback) {
+  console.log('sendAxiosMultipartRequest의 요청 URL: ', url);
+  console.log('sendAxiosMultipartRequest의 요청 데이터: ', formData);
+  const axiosConfig = {
+    timeout: 5000,
+    url: url,
+    method: 'POST',
+    data: formData,
+    headers: {
+      'Content-Type': 'multipart/form-data',
+    }
+  };
+  axios(axiosConfig).then(successCallback).catch(errorCallback);
 }
 
 
 export function dateFormatParse(date) {
-    const year = date.getFullYear();
-    const month = String(date.getMonth() + 1).padStart(2, '0');
-    const day = String(date.getDate()).padStart(2, '0');
-    const hours = String(date.getHours()).padStart(2, '0');
-    const minutes = String(date.getMinutes()).padStart(2, '0');
-    const seconds = String(date.getSeconds()).padStart(2, '0');
-    return `${year}/${month}/${day} ${hours}:${minutes}:${seconds}`;
+  const year = date.getFullYear();
+  const month = String(date.getMonth() + 1).padStart(2, '0');
+  const day = String(date.getDate()).padStart(2, '0');
+  const hours = String(date.getHours()).padStart(2, '0');
+  const minutes = String(date.getMinutes()).padStart(2, '0');
+  const seconds = String(date.getSeconds()).padStart(2, '0');
+  return `${year}/${month}/${day} ${hours}:${minutes}:${seconds}`;
 }
 
 export function handleInputChange(e, index, state, stateSetter) {
-    let {name, value} = e.target;
-    let deepCopyState = [...state];
-    deepCopyState[index] = {
-        ...deepCopyState[index],
-        [name]: value,
-    };
-    stateSetter(deepCopyState);
+  let { name, value } = e.target;
+  let deepCopyState = [...state];
+  deepCopyState[index] = {
+    ...deepCopyState[index],
+    [name]: value,
+  };
+  stateSetter(deepCopyState);
 }
 
 export function scrollMoveTop() {
-    window.scrollTo({top: 0, behavior: 'smooth'});
+  window.scrollTo({ top: 0, behavior: 'smooth' });
 }
 
 // 현재 일 기준 가게 시작, 종료 시간
 export function startEndTimeInfo(store) {
-    const currentDate = new Date();
-    const dayOfWeek = currentDate.getDay();
+  const currentDate = new Date();
+  const dayOfWeek = currentDate.getDay();
 
-    let startTime;
-    let endTime;
-    if (dayOfWeek === 0 || dayOfWeek === 6) { // 주말이면
-        startTime = store.weekendStartTime;
-        endTime = store.weekendEndTime;
-    } else { // 평일이면
-        startTime = store.weekdayStartTime;
-        endTime = store.weekdayEndTime;
-    }
+  let startTime;
+  let endTime;
+  if (dayOfWeek === 0 || dayOfWeek === 6) { // 주말이면
+    startTime = store.weekendStartTime;
+    endTime = store.weekendEndTime;
+  } else { // 평일이면
+    startTime = store.weekdayStartTime;
+    endTime = store.weekdayEndTime;
+  }
 
-    return [startTime, endTime];
+  return [startTime, endTime];
 }
 
 
 // 가게 시작시간, 종료시간 기준으로 영업중인지 확인
 export function isOpenNow(startTime, endTime) {
-    const currentDate = new Date();
-    const currentTimeMinutes = currentDate.getHours() * 60 + currentDate.getMinutes();
-    const start = startTime.split(':').map(Number);
-    const end = endTime.split(':').map(Number);
+  const currentDate = new Date();
+  const currentTimeMinutes = currentDate.getHours() * 60 + currentDate.getMinutes();
+  const start = startTime.split(':').map(Number);
+  const end = endTime.split(':').map(Number);
 
-    const startTimeMinutes = start[0] * 60 + start[1];
-    const endTimeMinutes = end[0] * 60 + end[1];
+  const startTimeMinutes = start[0] * 60 + start[1];
+  const endTimeMinutes = end[0] * 60 + end[1];
 
-    // O: 마감 C: 영업중
-    return currentTimeMinutes >= startTimeMinutes && currentTimeMinutes <= endTimeMinutes ? 'O' : 'C';
+  // O: 마감 C: 영업중
+  return currentTimeMinutes >= startTimeMinutes && currentTimeMinutes <= endTimeMinutes ? 'O' : 'C';
 }
 
 
 // 리스트(요소여러 개)에 이벤트 추가
 export function addEventToElements(eventName, eventFunction, elements) {
-    for (let element of elements) {
-        element.addEventListener(eventName, eventFunction);
-    }
+  for (let element of elements) {
+    element.addEventListener(eventName, eventFunction);
+  }
 }
 
 // 리스트(요소여러 개)에 이벤트 제거
 export function removeEventToElements(eventName, eventFunction, elements) {
-    for (let element of elements) {
-        element.removeEventListener(eventName, eventFunction);
-    }
+  for (let element of elements) {
+    element.removeEventListener(eventName, eventFunction);
+  }
 }
